requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
opencv-python>=4.8.0

# Metrics and evaluation
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Progress bars
tqdm>=4.65.0

# Logging (standard library, but listed for reference)
# logging

# Note: Install PyTorch with CPU support using:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
